requirements.txt
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
dill==0.3.5.1
idna==3.3
mpmath==1.2.1
ntlm-auth==1.5.0
numpy==1.23.2
pbr==5.10.0
Pillow==9.2.0
ply==3.11
psutil==5.9.2
pycparser==2.21
python-dateutil==2.8.2
qiskit==0.37.2
qiskit-aer==0.10.4
qiskit-ibmq-provider==0.19.2
qiskit-terra==0.21.2
requests==2.28.1
requests-ntlm==1.1.0
retworkx==0.11.0
scipy==1.9.1
six==1.16.0
stevedore==4.0.0
symengine==0.9.2
sympy==1.11.1
tweedledum==1.1.1
urllib3==1.26.12
websocket-client==1.4.1
websockets==10.3
